clean up and naming
