+++ demo/app.js
@@ -14,7 +14,7 @@ import GroupDemo from "./components/group-demo";
 import VoronoiDemo from "./components/victory-voronoi-demo";
 import TooltipDemo from "./components/victory-tooltip-demo";
 import ZoomContainerDemo from "./components/victory-zoom-container-demo";
-import RangeContainerDemo from "./components/victory-range-container-demo";
+import BrushContainerDemo from "./components/victory-brush-container-demo";
 import ZoomDemo from "./components/victory-zoom-demo";
 import AnimationDemo from "./components/animation-demo";
 import SelectionDemo from "./components/selection-demo";
@@ -45,7 +45,7 @@ const App = React.createClass({
           <li><Link to="/voronoi">Victory Voronoi Demo</Link></li>
           <li><Link to="/tooltip">Victory Tooltip Demo</Link></li>
           <li><Link to="/zoom-container">Victory Zoom Container Demo</Link></li>
-          <li><Link to="/range-container">Victory Range Container Demo</Link></li>
+          <li><Link to="/brush-container">Victory Brush Container Demo</Link></li>
           <li><Link to="/zoom">Victory Zoom Demo</Link></li>
           <li><Link to="/animation">Animation Demo</Link></li>
           <li><Link to="/selection">Selection Demo</Link></li>
@@ -72,7 +72,7 @@ ReactDOM.render((
       <Route path="voronoi" component={VoronoiDemo}/>
       <Route path="tooltip" component={TooltipDemo}/>
       <Route path="zoom-container" component={ZoomContainerDemo}/>
-      <Route path="range-container" component={RangeContainerDemo}/>
+      <Route path="brush-container" component={BrushContainerDemo}/>
       <Route path="zoom" component={ZoomDemo}/>
       <Route path="animation" component={AnimationDemo}/>
       <Route path="selection" component={SelectionDemo}/>
