--- src/components/containers/zoom-helper-methods.js
@@ -1,64 +1,0 @@
-import { Collection } from "victory-core";
-export default {
- /**
-   * Generates a new domain scaled by factor and constrained by the original domain.
-   * @param  {[Number, Number]} currentDomain  The domain to be scaled.
-   * @param  {[Number, Number]} originalDomain The original domain for the data set.
-   * @param  {Number}           factor         The delta to translate by
-   * @return {[Number, Number]}                The scale domain
-   */
-  scale: (currentDomain, originalDomain, factor) => {
-    const [fromBound, toBound] = originalDomain;
-    const [from, to] = currentDomain;
-    const range = Math.abs(from - to);
-    const midpoint = +from + (range / 2);
-    const newRange = (range * factor) / 2;
-    return [
-      Collection.getMaxValue([midpoint - newRange, fromBound]),
-      Collection.getMinValue([midpoint + newRange, toBound])
-    ];
-  },
-
-/**
- * Generate a new domain translated by the delta and constrained by the original domain.
- * @param  {[Number, Number]} currentDomain  The domain to be translated.
- * @param  {[Number, Number]} originalDomain The original domain for the data set.
- * @param  {Number}           delta          The delta to translate by
- * @return {[Number, Number]}                The translated domain
- */
-  pan: (currentDomain, originalDomain, delta) => {
-    const [fromCurrent, toCurrent] = currentDomain.map((val) => +val);
-    const [fromOriginal, toOriginal] = originalDomain.map((val) => +val);
-    const lowerBound = fromCurrent + delta;
-    const upperBound = toCurrent + delta;
-    let newDomain;
-    if (lowerBound > fromOriginal && upperBound < toOriginal) {
-      newDomain = [lowerBound, upperBound];
-    } else if (lowerBound < fromOriginal) { // Clamp to lower limit
-      const dx = toCurrent - fromCurrent;
-      newDomain = [fromOriginal, fromOriginal + dx];
-    } else if (upperBound > toOriginal) { // Clamp to upper limit
-      const dx = toCurrent - fromCurrent;
-      newDomain = [toOriginal - dx, toOriginal];
-    } else {
-      newDomain = currentDomain;
-    }
-    return Collection.containsDates(currentDomain) || Collection.containsDates(originalDomain) ?
-      newDomain.map((val) => new Date(val)) : newDomain;
-  },
-
-  getOriginalDomain: (scale) => {
-    return {
-      x: scale.x.domain(),
-      y: scale.y.domain()
-    };
-  },
-
-  getDomainScale: (domain, scale) => {
-    const {x: [from, to]} = domain;
-    const rangeX = scale.x.range();
-    const plottableWidth = Math.abs(rangeX[0] - rangeX[1]);
-    return plottableWidth / (to - from);
-  }
-};
-
